+++ quanatee/contrib/pentagon/api4twelve/api.go
@@ -156,13 +156,13 @@ func GetAggregates(
 					ohlcv.High[Epoch] = aggCrypto.PriceData[bar].High
 					ohlcv.Low[Epoch] = aggCrypto.PriceData[bar].Low
 					ohlcv.Close[Epoch] = aggCrypto.PriceData[bar].Close
-					ohlcv.Volume[Epoch] = 1.0
+					ohlcv.Volume[Epoch] = float32(1)
 					// Extra
 					ohlcv.HLC[Epoch] = (ohlcv.High[Epoch] + ohlcv.Low[Epoch] + ohlcv.Close[Epoch])/3
 					ohlcv.TVAL[Epoch] = ohlcv.HLC[Epoch] * ohlcv.Volume[Epoch]
 					ohlcv.Spread[Epoch] = ohlcv.High[Epoch] - ohlcv.Low[Epoch]
 					// Store Split Ratio
-					ohlcv.Split[Epoch] = 1.0
+					ohlcv.Split[Epoch] = float32(1)
 				}
 			}
 		} else if strings.Compare(marketType, "currency") == 0 {
@@ -184,13 +184,13 @@ func GetAggregates(
 					ohlcv.High[Epoch] = aggForex.PriceData[bar].High
 					ohlcv.Low[Epoch] = aggForex.PriceData[bar].Low
 					ohlcv.Close[Epoch] = aggForex.PriceData[bar].Close
-					ohlcv.Volume[Epoch] = 1.0
+					ohlcv.Volume[Epoch] = float32(1)
 					// Extra
 					ohlcv.HLC[Epoch] = (ohlcv.High[Epoch] + ohlcv.Low[Epoch] + ohlcv.Close[Epoch])/3
 					ohlcv.TVAL[Epoch] = ohlcv.HLC[Epoch] * ohlcv.Volume[Epoch]
 					ohlcv.Spread[Epoch] = ohlcv.High[Epoch] - ohlcv.Low[Epoch]
 					// Store Split Ratio
-					ohlcv.Split[Epoch] = 1.0
+					ohlcv.Split[Epoch] = float32(1)
 				}
 			}
 		} else if strings.Compare(marketType, "equity") == 0 {
@@ -217,14 +217,14 @@ func GetAggregates(
 					if aggEquity.PriceData[bar].Volume != 0 {
 						ohlcv.Volume[Epoch] = aggEquity.PriceData[bar].Volume
 					} else {
-						ohlcv.Volume[Epoch] = 1.0
+						ohlcv.Volume[Epoch] = float32(1)
 					}
 					// Extra
 					ohlcv.HLC[Epoch] = (ohlcv.High[Epoch] + ohlcv.Low[Epoch] + ohlcv.Close[Epoch])/3
 					ohlcv.TVAL[Epoch] = ohlcv.HLC[Epoch] * ohlcv.Volume[Epoch]
 					ohlcv.Spread[Epoch] = ohlcv.High[Epoch] - ohlcv.Low[Epoch]
 					// Store Split Ratio
-					ohlcv.Split[Epoch] = 1.0
+					ohlcv.Split[Epoch] = float32(1)
 				}
 			}
 		}
